+++ Editor/AssemblyInfo.cs
@@ -7,3 +7,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.Addressables.Tests")]
 [assembly: InternalsVisibleTo("PerformanceTests.Editor")]
 [assembly: InternalsVisibleTo("Unity.Localization.Editor")]
+[assembly: InternalsVisibleTo("Unity.Addressables.Android.Editor")]