+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.21.12] - 2023-05-04
+- Fixed issue with a potential race conditon when calling GetAssetBundle to load local files using WaitForCompletion
+- Fixed compiler error when using the profiler module on a noncaching platform.
+- Fixed issue where stack overflow occurs for http requests on Unity 2022.1+ when insecure requests are disallowed.
+- Added groups field to Inspector and move assets popup.
+- Fixed issue where sprite atlas cannot be assigned to an AssetReferenceAtlasedSprite.
+
 ## [1.21.10] - 2023-04-10
 - Fixed Addressables profiler not correctly displaying loaded Asset subObjects where a Asset was loaded due to a direct reference to one of its subObjects.
 - Fixed an issue where bundles built with the Append Hash to FIlename bundle naming option had strange interactions with the Prevent Updates option